requirements.txt
tqdm==4.49.0
numpy==1.18.5
pandas==1.1.2
tensorflow==2.3.0
matplotlib==3.3.2
pyenchant==3.1.1
scikit_learn==0.23.2
